requirements.txt
absl-py==2.2.1
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dm-haiku==0.0.13
docker-pycreds==0.4.0
equinox==0.11.12
etils==1.12.2
Farama-Notifications==0.0.4
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
gymnasium==1.1.1
idna==3.10
jax==0.5.3
jaxlib==0.5.3
jaxtyping==0.3.0
jmp==0.0.4
kiwisolver==1.4.8
matplotlib==3.10.1
ml_dtypes==0.5.1
mpmath==1.3.0
numpy==2.2.4
opt_einsum==3.4.0
optax==0.2.4
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
sentry-sdk==2.24.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
tabulate==0.9.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
wadler_lindig==0.1.4
wandb==0.19.8
